requirements.txt
Django~=2.0.6
djangorestframework~=3.8.2
pytest~=3.8.2